:hammer: move reducers
