+++ src/gui/component.container.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
-import { getComponentProps } from './redux/props'
-import { state as stateStore } from './redux'
+import { getComponentProps } from './redux/component/props'
+import { state as stateStore } from './redux/component'
 import Component from './component'
 
 const mapState = state => ({ ...getComponentProps(state), state: stateStore.get()(state) })
