+++ src/gui/redux/component/index.js
@@ -1,0 +1,2 @@
+export * from './component'
+export { default } from './component'
