+++ src/gui/redux/component/props/index.js
