+++ src/gui/redux/docgen/docgen.selectors.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect'
-import { docgen } from '../store'
+import { docgen } from '../reducers'
 
 const getProps = state => docgen.get()(state).props
 export const getPropsKeys = createSelector(getProps, p => Object.keys(p))
