+++ src/gui/redux/index.js
@@ -1,2 +1,1 @@
-export * from './store'
 export { default } from './store'
